+++ Sources/JESideMenuController/LayoutBuilder/SlideOutInlineLayoutBuilder.swift
@@ -13,6 +13,11 @@ import UIKit
  */
 struct SlideOutInlineLayoutBuilder: LayoutBuilding {
 
+    // MARK: - Protocol Properties
+
+    let spacing: CGFloat
+    let ipadWidth: CGFloat
+
     // MARK: - Internal Properties
 
     /// A containerView that hosts the content of the menu.