+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 ### Added
 - Documenting changes in CHANGELOG file.
 - Increased code coverage.
+- Added functionality to adjust the side menu configuration. (spacing, iPad menu width, drop shadow image etc.)
 
 ### Fixed
 - Rewording of description and fix typos in README.